Updated Howler.js with the playbackRate property in the Howl object
I have added the playbackRate property in the Howl object. I found myself extending Howler with this since I needed it in my game, and because it isn't that much work but a great feature that will hopefully help other people as well,  I thought I'd submit it here in a Pull Request.

The playbackRate is a WebAudio API-only property that enables you to play a sound at a different speed and pitch(useful for games) than normal. It even allows you to play a sound in reverse.
Values are: 1.0 (default) for normal speed, 0.5 for half speed, 2.0 for double, -1.0 for reverse, and so on.