Push all method calls in a queue before load
Previously, when a method is called before the sound is loaded,
the method call would be added to an event queue, but NEVER
removed again. Consequently, when the sound is played again, the
queued method is called again.

Here is an example that demonstrates the consequences:
var h = new Howl({urls:['http://goldfirestudios.com/proj/howlerjs/sound.mp3']})
h.volume(1);
h.on('load', function() {
    h.volume(0);
    // Sound should be silent, but instead it is at volume 1!
    h.play();
})

This patch replaces the erroneous "on('load'," implementation with
new internal methods. When methods are called before the sound is
loaded, the methods are pushed on a queue. When the sound is ready,
the queue is drained in FIFO order.

Another way to fix the bug is by implementing an once() method which
adds an event listener to an event queue, and removes the listener
after calling the method one time.