_clearEndTimer does not accidentally clear first timer
Why?
When the soundId passed to _clearEndTimer() does not exist, the first timer in the _onendTimer array will be cleared, instead it should do nothing. This situation will occur when playing several overlapping instances of the same sound, the "end" event will not fire for all instances.

How?
Start index at -1 instead of 0, timer will not exist if not found in the loop.