#1295: 'playing' method behavior change
The current 'playing' method behavior is a bit misleading, since it will only check the first instance of the sound group if an id is not passed. Checking the whole sound group would be far more useful, imho. My proposed change is when the user calls the method without passing an id, the whole group will be loop-checked. If at least one of the sound instances is still playing, the method will return true.