Use the MediaError message attribute if available
The `message` attribute on the `MediaError` object returned by
`HTMLMediaElement.error` contains specific diagnostic details about the
error if available, or an empty string otherwise.

If this message is available then it will be passed with the
`loaderror`, otherwise the code will be passed as before.

See https://html.spec.whatwg.org/multipage/media.html#error-codes.